Add authentication (still WIP) and upvote functionalities, as well the latest data structure
Authentication is still WIP, it works with Postman but not with the React app.
Upvote functionalities allow to upvote or remove upvote for an article once the user is logged in. The upvote is linked to its author and the article. When upvote is removed by the user, and is not removed from the database, but deactivated.
Data structure is now using populate to get an article's author, and virtual populates to get an article's number of upvotes, and an author articles posted or upvotes.