Add the docker functionalities for the react app and backend separately, still WIP.
Added the docker functionalities for both the react app and the backend. While the app works when launched on the local machine, it doesn't work when launched in the docker containers. The react app gives an error that needs to be solved. Also, puppeteer was installed in the backend docker container, but remains to be tested.