Add the Docker multistage for react app, and a first draft of the CD GitHub actions
Implemented the Docker multistage with nginx for the react app. This is still work in progress as it was giving a 403 error when running the container in local and trying to access the site. ALso implemented a first draft of the continous deployment GitHub actions for code pushed to main branch.