Add authentication on websocket.
Add a first http request to the server to get a token. The token is generated by the server, saved on the user in the database and sent to the front end. Then the frontend sends the token as a message when openning the websocket connection, the token is checked against the token saved on the user in the database.